update mixed workspace test, split for abs and s3
